// src/interfaces/appInterfaces.ts
// Generated by https://quicktype.io

export interface LoginResponse {
    usuario: Usuario;
    token: string;
}

export interface Usuario {
    rol: string;
    estado: boolean;
    google: boolean;
    nombre: string;
    correo: string;
    uid: string;
    img?: string;
}


// Interfaces propias

export interface LoginData {
    correo: string,
    password: string
}

export interface RegisterData {
    nombre: string
    correo: string
    password: string
}


export interface CreateProductData {
    nombre: string,
    categoria: string
}

// * Productos Interfaces

export interface GetProductsResponse {
    total: number;
    productos: Producto[];
}

export interface Producto {
    precio: number;
    _id: string;
    nombre: string;
    categoria: Categoria;
    usuario: Categoria;
    img?: string;
}

// export interface Categoria {
//     _id: string;
//     nombre: string;
// }



// * Categorias Interfaces

export interface GetCategoriasResponse {
    total: number;
    categorias: Categoria[];
}

export interface Categoria {
    _id: string;
    nombre: string;
    usuario: CreadorPor;
}

export interface CreadorPor {
    _id: string;
    nombre: string;
}